Move save function to where its used
